Former-commit-id: 8ee3557b04ad9054e8e0c0e0671efa6f4e3abdad

+++ rpms/awips2.cave/Installer.cave-wrapper/component.spec
@@ -138,12 +138,30 @@ if [ $? -ne 0 ]; then
 fi
 popd > /dev/null 2>&1
 
+pushd . > /dev/null 2>&1
+cd %{_baseline_workspace}/viz.updater
+if [ ! -d bin ]; then
+   mkdir -p bin
+fi
+/awips2/ant/bin/ant -Ddest.dir=${RPM_BUILD_ROOT}/awips2/cave \
+   -Declipse.dir=%{_uframe_eclipse} -Dbaseline.dir=%{_baseline_workspace} -f build.xml
+if [ $? -ne 0 ]; then
+   exit 1
+fi
+popd > /dev/null 2>&1
+
 %pre
 %post
 # Remove the text-workstation autostart script if we have not been installed
 # on an xt workstation
 if [ ! "`hostname | cut -b 1-2`" = "xt" ]; then
    rm -f /etc/xdg/autostart/awips2-textws.desktop
+fi
+
+# relocate any localization files installed by awips2-cave
+/bin/bash /awips2/cave/relocateLocalization.sh
+if [ $? -ne 0 ]; then
+   exit 1
 fi
 
 %preun
@@ -168,6 +186,7 @@ rm -rf ${RPM_BUILD_ROOT}
 %dir /awips2/cave/caveEnvironment
 /awips2/cave/caveEnvironment/*
 /awips2/cave/*.sh
+/awips2/cave/VizUpdater.jar
 # not a noarch RPM due to the presence of the architecture-specific libraries.
 %dir /awips2/cave/lib%{_build_bits}
 /awips2/cave/lib%{_build_bits}/*
